requirements.txt
requests
beautifulsoup4
pandas
streamlit
psycopg2-binary
sqlalchemy
plotly